[ImportStock] Warehouse import stock, fix remove bug on orphan

+++ src/main/java/com/wifosell/zeus/service/impl/batch_process/warehouse/ImportStockRow.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 public class ImportStockRow {
     public String sku;
     public Integer quantity;
-    public Integer unitCost;
+    public BigDecimal unitCost;
 }
